requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
DiscordTools
agno
sqlalchemy
pgvector
psycopg[binary]
pypdf
google-generativeai
huggingface-hub
sentence-transformers
streamlit
plotly
streamlit
pandas
yfinance
plotly
requests
openai
numpy
matplotlib
statsmodels
scikit-learn
tensorflow
google-genai